LayoutTests/imported/w3c/web-platform-tests/reporting/document-reporting-destroy-after-document-close.https.sub-expected.txt



FAIL No more reports received after navigation to same document without endpoint header assert_equals: expected 1 but got 0
